+++ packages/legend-query-builder/src/components/fetch-structure/QueryBuilderTDSPanel.tsx
@@ -30,7 +30,6 @@ import {
   PanelDropZone,
   DragPreviewLayer,
   useDragPreviewLayer,
-  OptionsIcon,
   PlusIcon,
   PanelContent,
   TrashIcon,
@@ -41,6 +40,7 @@ import {
   CustomSelectorInput,
   PanelEntryDropZonePlaceholder,
   FunctionIcon,
+  CogIcon,
 } from '@finos/legend-art';
 import {
   type QueryBuilderExplorerTreeDragSource,
@@ -952,6 +952,7 @@ export const QueryBuilderTDSPanel = observer(
     const clearAllProjectionColumns = (): void => {
       tdsState.checkBeforeClearingColumns(() => {
         tdsState.removeAllColumns();
+        tdsState.resultSetModifierState.reset();
       });
     };
 
@@ -1061,35 +1062,92 @@ export const QueryBuilderTDSPanel = observer(
     return (
       <PanelContent>
         <div className="query-builder__projection__toolbar">
-          <button
-            className="panel__header__action"
-            onClick={openResultSetModifierEditor}
-            tabIndex={-1}
-            title="Configure result set modifiers..."
-          >
-            <OptionsIcon className="query-builder__icon query-builder__icon__query-option" />
-          </button>
-          <button
-            className="panel__header__action"
-            disabled={isEmpty}
-            onClick={clearAllProjectionColumns}
-            tabIndex={-1}
-            title={
-              isEmpty
-                ? 'No projection columns to clear'
-                : 'Clear all projection columns'
+          <div
+            className="query-builder__projection__result-modifier-prompt"
+            data-testid={
+              QUERY_BUILDER_TEST_ID.QUERY_BUILDER_TDS_RESULT_MODIFIER_PROMPT
             }
           >
-            <TrashIcon className="query-builder__icon query-builder__icon__query-option--small" />
-          </button>
-          <button
-            className="panel__header__action"
-            onClick={addNewBlankDerivation}
-            tabIndex={-1}
-            title="Add a new derivation"
-          >
-            <PlusIcon />
-          </button>
+            <div className="query-builder__projection__result-modifier-prompt__header">
+              <button
+                className="query-builder__projection__result-modifier-prompt__header__label"
+                onClick={openResultSetModifierEditor}
+                title="Configure result set modifiers..."
+              >
+                <CogIcon className="query-builder__projection__result-modifier-prompt__header__label__icon" />
+                <div className="query-builder__projection__result-modifier-prompt__header__label__title">
+                  Query Options
+                </div>
+              </button>
+            </div>
+            {tdsState.resultSetModifierState.limit && (
+              <div className="query-builder__projection__result-modifier-prompt__group">
+                <div className="query-builder__projection__result-modifier-prompt__group__label">
+                  Max Rows
+                </div>
+                <div
+                  className="query-builder__projection__result-modifier-prompt__group__content"
+                  onClick={openResultSetModifierEditor}
+                >
+                  {tdsState.resultSetModifierState.limit}
+                </div>
+              </div>
+            )}
+            {tdsState.resultSetModifierState.distinct && (
+              <div className="query-builder__projection__result-modifier-prompt__group">
+                <div className="query-builder__projection__result-modifier-prompt__group__label">
+                  Eliminate Duplicate Rows
+                </div>
+                <div
+                  className="query-builder__projection__result-modifier-prompt__group__content"
+                  onClick={openResultSetModifierEditor}
+                >
+                  Yes
+                </div>
+              </div>
+            )}
+            {tdsState.resultSetModifierState.sortColumns.length > 0 && (
+              <div className="query-builder__projection__result-modifier-prompt__group">
+                <div className="query-builder__projection__result-modifier-prompt__group__label">
+                  Sort
+                </div>
+                {tdsState.resultSetModifierState.sortColumns.map(
+                  (columnState) => (
+                    <div
+                      className="query-builder__projection__result-modifier-prompt__group__content"
+                      key={columnState.columnState.uuid}
+                      onClick={openResultSetModifierEditor}
+                    >
+                      {`${columnState.columnState.columnName} ${columnState.sortType}`}
+                    </div>
+                  ),
+                )}
+              </div>
+            )}
+          </div>
+          <div className="query-builder__projection__toolbar__actions">
+            <button
+              className="panel__header__action"
+              disabled={isEmpty}
+              onClick={clearAllProjectionColumns}
+              tabIndex={-1}
+              title={
+                isEmpty
+                  ? 'No projection columns to clear'
+                  : 'Clear all projection columns'
+              }
+            >
+              <TrashIcon className="query-builder__icon query-builder__icon__query-option--small" />
+            </button>
+            <button
+              className="panel__header__action"
+              onClick={addNewBlankDerivation}
+              tabIndex={-1}
+              title="Add a new derivation"
+            >
+              <PlusIcon />
+            </button>
+          </div>
         </div>
         <div className="query-builder__projection__content">
           <PanelDropZone