+++ packages/legend-application-query/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-application-query
 
+## 13.1.7
+
+## 13.1.6
+
 ## 13.1.5
 
 ## 13.1.4