+++ packages/legend-extension-assortment/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-assortment
 
+## 0.0.60
+
+## 0.0.59
+
 ## 0.0.58
 
 ## 0.0.57