+++ packages/legend-extension-dsl-data-space/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-data-space
 
+## 10.1.26
+
+## 10.1.25
+
 ## 10.1.24
 
 ## 10.1.23