+++ packages/legend-extension-dsl-mastery/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @finos/legend-extension-dsl-mastery
 
+## 3.0.83
+
+### Patch Changes
+
+- [#2613](https://github.com/finos/legend-studio/pull/2613) [`03f58c386`](https://github.com/finos/legend-studio/commit/03f58c386dc927f5c6eaf1216232f5a5f22b7ec7) ([@sahil37](https://github.com/sahil37)) - adding additional fields to mastery spec
+
+## 3.0.82
+
 ## 3.0.81
 
 ### Patch Changes