+++ packages/legend-extension-dsl-persistence/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-persistence
 
+## 5.0.82
+
+## 5.0.81
+
 ## 5.0.80
 
 ## 5.0.79