+++ packages/legend-extension-dsl-text/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-dsl-text
 
+## 6.0.82
+
+## 6.0.81
+
 ## 6.0.80
 
 ## 6.0.79