+++ packages/legend-extension-store-relational/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @finos/legend-extension-store-relational
 
+## 0.0.154
+
+## 0.0.153
+
 ## 0.0.152
 
 ## 0.0.151